Templates Should write to a buffer
When templates write directly to a `http.ResponseWriter`, there is
a risk of an error. When there is an error, `WriteHeader()` is
accidentally called twice with conflicting status codes, and an
incomplete page is rendered. Instead, we ensure that the templates
do not error before we write to the `http.ResponseWriter`.